DAOS-17941 client: re-organize share memory related header files
